+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * improve aura handler icon
 * fix ember/pressure-air ui progress bar
+* add machine sound (you can set sound for each machine state)